concepts: Reorganize implementations
non-trivial concepts are now split into two parts, each within their own
comptime block: validation and auditing.

The former checks that the inputs for a given concept are correct. The
latter checks that the requirements of a given concept are satisfied.

Each one is in a comptime block because otherwise, Zig's lazy
compilation would show compile errors that clutter up the output.